Move secondary_code field
